+++ index.js
@@ -1,9 +1,9 @@
-const SslCommerzPayment = require("./api/payment-controller");
+import SslCommerzPaymentBase from "./api/payment-controller.js";
 
-class SSLCommerzPayment extends SslCommerzPayment {
-    constructor(store_id, store_password, live = false) {
-        super(store_id, store_password, live);
-    }
+class SSLCommerzPayment extends SslCommerzPaymentBase {
+	constructor(store_id, store_password, live = false) {
+		super(store_id, store_password, live);
+	}
 }
 
-module.exports = SSLCommerzPayment;
+export default SSLCommerzPayment;