+++ CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ### Major Changes
 
-- Updated node-fetch, form-data dependencies to work with latest version of nodejs application like next.js 14.X.X
+- Updated node-fetch, form-data dependencies and ESM applied to work with latest version of nodejs application like next.js 14.X.X